// src/types/Airline.ts
export default interface Airline {
    id: string;
    fleetAverageAge: string;
    airlineId: string;
    callsign: string;
    hubCode: string;
    iataCode: string;
    icaoCode: string;
    countryIso2: string;
    dateFounded: string;
    iataPrefixAccounting: string;
    airlineName: string;
    countryName: string;
    fleetSize: string;
    status: string;
    type: string;
}